firmware/requirements.txt
requests
rpi_ws281x
adafruit-circuitpython-neopixel
